refactor(web): move the org-creation form into its own component
